Day 6 works.
There is a pattern emerging of input files where data for each object is spread across multiple lines, which are then separated by blank lines. We saw this in Day 4 and we see it again in Day 6. I created a data type to handle this processing, rewrote Day 4 to use it, then applied it on Day 6.